Move some dims utils to common (#2223)
